pyproject.toml
[tool.poetry]
name = "pyboost"
version = "0.1.0"
description = "Ferramenta de linha de comando (CLI) que visa acelerar o desenvolvimento de aplicativos em Python."
authors = ["Nícolas Albuquerque Ramos <nicolasalbuquerque581@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.scripts]
pyboost = "pyboost.cli:app"

[tool.black]
line-length = 79
single-quote = true

[tool.poetry.dependencies]
python = "^3.10"
typer-cli = "^0.0.13"
rich = "^13.3.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
